requirements.txt
streamlit
PyPDF2
openai
python-dotenv
langchain-community
faiss-cpu
tiktoken

